requirements.txt
streamlit
google-genai
langchain
langchain_groq
